miedo

+++ bocana/src/modelo/OfertaDao.java
@@ -30,7 +30,7 @@ public class OfertaDao {
             ps = con.prepareStatement(sql);
 
             ps.setInt(1, o.getIdOfertaEspecial());
-            ps.setInt(2, o.getDescuento());
+            ps.setDouble(2, o.getDescuento());
             ps.setDate(3, (Date) o.getFechaInicio());
             ps.setDate(4, (Date) o.getFechaFin());
